@@ -5,7 +5,7 @@ using Xunit.Abstractions;
 namespace Respawn.DatabaseTests
 {
     using System;
-    using System.Data.SqlClient;
+    using Microsoft.Data.SqlClient;
     using System.Linq;
     using NPoco;
     using Shouldly;
