@@ -7,7 +7,7 @@ namespace Respawn
 	using System;
 	using System.Collections.Generic;
 	using System.Data.Common;
-	using System.Data.SqlClient;
+	using Microsoft.Data.SqlClient;
 	using System.Linq;
 	using System.Threading.Tasks;
 
